Add get-weather script
This adds a script used by xmobar to get the weather. It reads a partial
JavaScript file provided by Environment Canada and returns a string with
an icon and the temperature (color coded with a threshold). Passing the
--url argument opens a new Google Chrome tab with more detailed weather
information.

The script needs to handle a greater number of weather conditions (rain,
etc.), windchill display, and weather warning/watch/statement display. I
may also add the forecast in at some point, but I'm not sure how I could
display that nicely.

I'm currently using the Calgary International Airport station, but the
Calgary Olympic Park station is closer to my home. Unfortunately, that
station does not report with as much detail, so I'll stick with the
airport for now.