Add get-config script
This commit adds a script that reads my ~/.config/config.json file for
config information. I've added that script to the path using my utility
manager, and am now using it in some of my xmobar scripts to get
information. This should simplify things in the future when I want to
update those configuration items with shortcuts or scripts.

I'd like to expand the functionality of the get-config script to also
change my config, but at that point it will probably need a new name
(perhaps config-manager.sh).